+++ src/lib.rs
@@ -149,7 +149,7 @@ pub use slang::{
 pub use task_graph::NodeAccess;
 #[cfg(feature = "tensor")]
 pub use tensor::{
-    GoldyTensorLayout, ScatterMode, Tensor, TensorContext, TensorDType, TensorLayout, TensorRecorder, TensorScalar,
+    GoldyTensorLayout, ScatterMode, Tensor, TensorDType, TensorKernels, TensorLayout, TensorRecorder, TensorScalar,
     TensorShape, TensorView, MAX_TENSOR_RANK,
 };
 pub use texture::TextureCopyFootprint;